+++ sdk/core/azure-core/CMakeLists.txt
@@ -52,6 +52,7 @@ set(
     inc/azure/core/http/transport.hpp
     inc/azure/core/internal/contract.hpp
     inc/azure/core/internal/hkeyholder.hpp
+    inc/azure/core/internal/http/null_body_stream.hpp
     inc/azure/core/internal/http/pipeline.hpp
     inc/azure/core/internal/json_serializable.hpp
     inc/azure/core/internal/json.hpp
